+++ lib/whois/record/parser/whois.nic.tel.rb
@@ -7,68 +7,19 @@
 #++
 
 
-require 'whois/record/parser/base'
+require 'whois/record/parser/base_shared2'
 
 
 module Whois
   class Record
     class Parser
 
-      #
-      # = whois.nic.tel parser
-      #
       # Parser for the whois.nic.tel server.
       #
-      # NOTE: This parser is just a stub and provides only a few basic methods
-      # to check for domain availability and get domain status.
-      # Please consider to contribute implementing missing methods.
-      # See WhoisNicIt parser for an explanation of all available methods
-      # and examples.
+      # @see Whois::Record::Parser::Example
+      #   The Example parser for the list of all available methods.
       #
-      class WhoisNicTel < Base
-
-        property_supported :status do
-          if available?
-            :available
-          else
-            :registered
-          end
-        end
-
-        property_supported :available? do
-          !!(content_for_scanner =~ /Not found:/)
-        end
-
-        property_supported :registered? do
-          !available?
-        end
-
-
-        property_supported :created_on do
-          if content_for_scanner =~ /Domain Registration Date:\s+(.*)\n/
-            Time.parse($1)
-          end
-        end
-
-        property_supported :updated_on do
-          if content_for_scanner =~ /Domain Last Updated Date:\s+(.*)\n/
-            Time.parse($1)
-          end
-        end
-
-        property_supported :expires_on do
-          if content_for_scanner =~ /Domain Expiration Date:\s+(.*)\n/
-            Time.parse($1)
-          end
-        end
-
-
-        property_supported :nameservers do
-          content_for_scanner.scan(/Name Server:\s+(.+)\n/).flatten.map do |name|
-            Record::Nameserver.new(:name => name.downcase)
-          end
-        end
-
+      class WhoisNicTel < BaseShared2
       end
 
     end