+++ spec/whois/record/parser/responses/whois.nic.us/status_available_spec.rb
@@ -21,9 +21,19 @@ describe Whois::Record::Parser::WhoisNicUs, "status_available.expected" do
     described_class.new(part)
   end
 
+  describe "#domain" do
+    it do
+      subject.domain.should == "u34jedzcq.us"
+    end
+  end
+  describe "#domain_id" do
+    it do
+      subject.domain_id.should == nil
+    end
+  end
   describe "#status" do
     it do
-      subject.status.should == []
+      subject.status.should == nil
     end
   end
   describe "#available?" do
@@ -49,6 +59,29 @@ describe Whois::Record::Parser::WhoisNicUs, "status_available.expected" do
   describe "#expires_on" do
     it do
       subject.expires_on.should == nil
+    end
+  end
+  describe "#registrar" do
+    it do
+      subject.registrar.should == nil
+    end
+  end
+  describe "#registrant_contacts" do
+    it do
+      subject.registrant_contacts.should be_a(Array)
+      subject.registrant_contacts.should == []
+    end
+  end
+  describe "#admin_contacts" do
+    it do
+      subject.admin_contacts.should be_a(Array)
+      subject.admin_contacts.should == []
+    end
+  end
+  describe "#technical_contacts" do
+    it do
+      subject.technical_contacts.should be_a(Array)
+      subject.technical_contacts.should == []
     end
   end
   describe "#nameservers" do